refactor: split controller for few files. add router
