+++ .env-example
@@ -1,2 +1,1 @@
-HTTP_PORT = 8000
-WS_PORT = 8080
+PORT = 8000
