+++ src/ws_server/controllers/index.ts
@@ -1,0 +1,3 @@
+export * from './mouseMoving';
+export * from './drawing';
+export * from './screenCapturing';
