+++ src/ws_server/server.ts
@@ -1,5 +1,5 @@
 import { AddressInfo, WebSocketServer } from 'ws';
-import { handler } from './controllers';
+import { handler } from './router';
 import { Server } from 'http';
 
 export function startServer(server: Server) {
@@ -9,11 +9,14 @@ export function startServer(server: Server) {
   console.log();
   wss.on('connection', ws => {
     ws.on('message', async message => {
-      console.log(message.toString());
-      const [command] = message.toString().split(' ');
-      const result = await handler(message);
-      ws.send(`${command} ${result ?? ''}`);
-
+      try {
+        console.log(message.toString());
+        const [command] = message.toString().split(' ');
+        const result = await handler(message);
+        ws.send(`${command} ${result ?? ''}`);
+      } catch (e) {
+        console.log('Error has occurred', e);
+      }
     });
   });
   console.log(`WS server started on the ${(wss.address() as AddressInfo).port} port`);
